Update README.md
Abstract 
The goal of this project is to create an interactive dashboard application to be used by marketing, product development & sales team of a telecom company. This application uses customer churn data and returns graphs and figures with insights that can lead to development of customized products and marketing strategies aimed towards building customer loyalty. It can also be used to develop a sales strategy that can bring the right product to the right customer at the right time. The project also predicts churn rate using various prediction models, the best of which may be used by the company in future for predicting the churn rate.
Business Problem
Churn Rate is one of the major costs for any telecommunication company. As per the European Business review, even a churn rate under 2% can lead to loses of 65$ million per month from churn only. Objective of our project is to provide an interactive dashboard to understand the reasons (variables) behind churn rate in different customer segments and generate predictions to minimize the churn rate in the future.
Analytics Problem
As analysts, our goal is to:
•	Gather insights from available data about customer behavior, demographics and usage pattern
•	Utilize these insights to predict customers who are likely to churn (classification problem)
•	We will use AUC curve as a measure of validity for our prediction models
Data
This project uses an IBM sample set called Telecom Customer Churn available for the purpose of practicing data analysis on a ‘real-world’ type of business problem. The dataset is meant to simulate a typical dataset of a Telecommunications company. This application can be applied to real datasets.
The dataset contains 7043 observations (rows) and 22 variables (columns) that contain information about customer demographics (gender, senior citizenship status, children, and marital status), services they signed up for (phone line, multiple lines, online security, online backup, device protection, tech support, streaming TV and Movies), account information (type of contract, payment type, paperless billing, monthly and total charges), and churn (which customers left within the past month when the data was collected).
Methodology Selection
To come up with better marketing strategies and target customers that would stay, an organization needs to perform both descriptive and predictive analytics to understand the churn rate. As R provides a comprehensive package for performing both descriptive and predictive analytics, it is the most suitable choice for our project.
The raw dataset is imported and cleaned in an R file in the "data" folder. The dashboard application is created using the R Shiny app package and makes use of the shinydashboard library for the sidebar layout. The ggplot2 library is used for creating graphs, while gridExtra is used to layout graphs side-by-side in some cases. The dplyr library filter function is used to filter the data. And finally, h2o is used for applying machine learning models on the data.
Model Building
For prediction modeling, we attempted 3 different machine learning models: Gradient Boosting, Random Forest, Deep Learning.
Functionality
Our app is designed to display both descriptive and predictive analytics for the data governing the churn rate. Using the app, the client can extract the following insights:
1.	Customer Demographics – Count, Contract type, Payment methods, charges, tenure, service
2.	Category of customers – Senior citizen, gender, relationship status and number of dependents
3.	Variables/Factors impacting the churn rate 
4.	 Model accuracy using the confusion matrices for the models used for prediction of churn
GUI Design and Functionality
The design of the application makes use of the shinydashboard library. The dashboard has a sidebar with drop-down menus to select demographic filters, which can be toggled to expose only the main panel.
The color theme is consistent throughout the entire dashboard, with blue representing all customers in the target demographic, green for customers that stayed and orange for customers that left. These colors are also used in the graphs to show the distribution of the of customers that stayed and left together. For the confusion matrices, “green” is used to represent the maximum value where as “orange” represents the minimum value. 
Conclusions
From our analysis, it can be concluded that the following factors are critical when predicting churn:
1.	Contract: Among those who churned, monthly contract customers were highest in number
2.	Tenure: Customers with longer tenure with the company tended to churn less
3.	Internet Service: Customers with no internet service tended to churn less
4.	Monthly Charges: Customers with lower monthly charges churned less
5.	Total Charges: Customers with lower Total Charges tended to churn
Gradient Boosting model performed the best and can be used for analysis in future.
References (if any)
Margaret Sant: https://github.com/santma/Telecommunications-Customer-Attrition-Dashboard